# improved runtime when cmd & file logging is off