--- a/+logging/clearLogger.m
+++ b/+logging/clearLogger.m
@@ -1,0 +1,5 @@
+function clearLogger(name)
+    [~, destructor] = logging.getLogger(name);
+    destructor();
+end
+