--- a/+logging/testspeed.m
+++ b/+logging/testspeed.m
@@ -6,10 +6,10 @@ function testspeed(logPath)
   L.setCommandWindowLevel(L.TRACE);
   L.setLogLevel(L.OFF);
   tic;
-  for i=1:1e2
+  for i=1:1e3
     L.trace('test');
   end
-  disp('1e2 logs when logging only to command window');
+  disp('1e3 logs when logging only to command window');
   toc;
 
   L.setCommandWindowLevel(L.OFF);
@@ -24,9 +24,9 @@ function testspeed(logPath)
   L.setCommandWindowLevel(L.OFF);
   L.setLogLevel(L.TRACE);
   tic;
-  for i=1:1e2
+  for i=1:1e3
     L.trace('test');
   end
-  disp('1e2 logs when logging to file');
+  disp('1e3 logs when logging to file');
   toc;
 end