Hi,

This pull request has,
- clean up handler in the getLogger function. 
- fixed the display of 2d char array message
- the code to avoid  `dbstack()` call when the command window logging and file logging is turned off. 

|	Task (1e3 lines)	   | previous elapsed time | current elapsed time |
|-----------------------------|-----------------------|----------------------|
| logging to command window | 3.363777		       | 3.133629
| logging is off		                   | **0.213782**		       | **0.011460**
| logging to file		           | 0.857700		       | 0.814813