@dpo 2017-04-21
Many thanks! Does your benchmark still hold?