@ragavsathish 2017-04-21
Please find the latest benchmarks

|	Task (1e3 lines)	   | previous elapsed time | current elapsed time |
|-----------------------------|-----------------------|----------------------|
| logging to command window | 1.452614		       |1.478851
| logging is off		                   | **0.105243**		       | **0.016133**
| logging to file		           | 0.637822		       | 0.621561
|logging to file when using function handle (Lazzy logging)  #8 | - | 0.720508